requirements.txt
torch
torchvision
streamlit
matplotlib
Pillow